Fix init issues
