+++ quasar.config.js
@@ -28,7 +28,7 @@ module.exports = configure(function (/* ctx */) {
     // app boot file (/src/boot)
     // --> boot files are part of "main.js"
     // https://v2.quasar.dev/quasar-cli-vite/boot-files
-    boot: ['extensions', 'i18n', 'mitt', 'dexie', 'init', 'checkForUpdate'],
+    boot: ['extensions', 'i18n', 'mitt', 'dexie', 'checkForUpdate'],
 
     // https://v2.quasar.dev/quasar-cli-vite/quasar-config-js#css
     css: ['app.scss'],
