+++ src/layouts/MainLayout.vue
@@ -120,7 +120,6 @@ import { useSettingsStore } from 'src/stores/settingsStore'
 import { Keyboard } from '@capacitor/keyboard'
 import { Platform } from 'quasar'
 import { useI18n } from 'vue-i18n'
-import { initSettingsLate } from 'src/scripts/initSettingsLate'
 
 const router = useRouter()
 const settingsStore = useSettingsStore()
@@ -189,7 +188,6 @@ function add() {
 }
 
 onMounted(() => {
-  initSettingsLate()
   if (Platform.is.mobile) {
     addKeyboardListeners()
   }
